+++ package.json
@@ -12,7 +12,7 @@
     "lint": "eslint ./src ./tests ./benchmark ./*.js",
     "lint:staged": "lint-staged",
     "prepublish": "npm run all",
-    "test": "cross-env BABEL_ENV=test mocha --require babel-core/register ./test"
+    "test": "cross-env BABEL_ENV=test jest ./test"
   },
   "author": {
     "name": "Oleg Slobodskoi",
@@ -44,7 +44,6 @@
     "babel-plugin-transform-es3-property-literals": "^6.22.0",
     "babel-preset-es2015": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
-    "chai": "^4.1.2",
     "cross-env": "^5.0.0",
     "es5-shim": "^4.3.1",
     "eslint": "^3.0.1",
@@ -53,9 +52,9 @@
     "eslint-plugin-import": "2.2.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.5.0",
+    "jest": "^23.4.1",
     "json-loader": "^0.5.4",
     "lint-staged": "^3.4.2",
-    "mocha": "^5.2.0",
     "pre-commit": "^1.2.0",
     "webpack": "^2.5.1"
   },
